+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Nylas Java SDK Changelog
 
 ### Unreleased
+* Made `NylasClient` and its methods open to enable mocking in tests
 * Added pagination support for folders
 
 ### [2.5.2] - Released 2024-12-02